Added CleanOutputDirectoryAction to delete the previous generated file before creating new ones.
